+++ google-chat-linux.sh
@@ -4,5 +4,5 @@ echo $SCRIPTPATH
 export NODE_PATH="$SCRIPTPATH/node_modules/.bin"
 export PATH=$NODE_PATH:$PATH
 
-${NODE_PATH}/electron "${SCRIPTPATH}/src/index.js"
+${NODE_PATH}/electron  "${SCRIPTPATH}/src/index.js" --trace-warnings